Refactor: Consolidate redundant test suites and cases
- Deleted redundant tests/unit/services.test.ts as coverage exists in services-specific test files.
- Reorganized tests/unit/websocketUtils.test.ts into the utils directory and updated imports.
- Merged server and client logger tests into a single tests/unit/utils/logger.test.ts using dynamic imports for environment isolation.
- Enhanced integration test helpers to support environment variable overrides, ensuring reliable rate-limit and connection-limit testing.
- Fixed linting errors and verified that both unit and integration test suites pass successfully.
